+++ Outline.Core/Packages/UnityFx.Outline/Editor/Scripts/OutlineSettingsEditor.cs
@@ -12,22 +12,166 @@ namespace UnityFx.Outline
 	[CustomEditor(typeof(OutlineSettings))]
 	public class OutlineSettingsEditor : Editor
 	{
-		private OutlineSettings _settings;
+		private const string _layerMaskPropName = "_outlineLayerMask";
+		private const string _settingsPropName = "_outlineSettings";
+		private const string _colorPropName = "_outlineColor";
+		private const string _widthPropName = "_outlineWidth";
+		private const string _intensityPropName = "_outlineIntensity";
+		private const string _cutoffPropName = "_outlineAlphaCutoff";
+		private const string _renderModePropName = "_outlineMode";
+
+		private static readonly GUIContent _layerMaskContent = new GUIContent("Outline Layer Mask", OutlineResources.OutlineLayerMaskTooltip);
+		private static readonly GUIContent _colorContent = new GUIContent("Color", "Outline color.");
+		private static readonly GUIContent _widthContent = new GUIContent("Width", "Outline width in pixels.");
+		private static readonly GUIContent _renderModeContent = new GUIContent("Render Flags", "Outline render flags. Multiple values can be selected at the same time.");
+		private static readonly GUIContent _intensityContent = new GUIContent("Blur Intensity", "Outline intensity value. It is only usable for blurred outlines.");
+		private static readonly GUIContent _cutoffContent = new GUIContent("Alpha Cutoff", "Outline alpha cutoff value. It is only usable when alpha testing is enabled and the material doesn't have _Cutoff property.");
 
 		public override void OnInspectorGUI()
 		{
-			EditorGUI.BeginChangeCheck();
-			OutlineEditorUtility.Render(_settings, _settings);
+			base.OnInspectorGUI();
 
-			if (EditorGUI.EndChangeCheck())
+			var colorProp = serializedObject.FindProperty(_colorPropName);
+			var widthProp = serializedObject.FindProperty(_widthPropName);
+			var intensityProp = serializedObject.FindProperty(_intensityPropName);
+			var cutoffProp = serializedObject.FindProperty(_cutoffPropName);
+			var renderModeProp = serializedObject.FindProperty(_renderModePropName);
+			var renderMode = (OutlineRenderFlags)renderModeProp.intValue;
+
+			EditorGUILayout.PropertyField(colorProp, _colorContent);
+			EditorGUILayout.PropertyField(widthProp, _widthContent);
+
+			//EditorGUILayout.PropertyField(renderModeProp, _renderModeContent);
+			renderModeProp.intValue = (int)(OutlineRenderFlags)EditorGUILayout.EnumFlagsField(_renderModeContent, renderMode);
+
+			if ((renderMode & OutlineRenderFlags.Blurred) != 0)
 			{
-				EditorUtility.SetDirty(_settings);
+				EditorGUILayout.PropertyField(intensityProp, _intensityContent);
+			}
+
+			if ((renderMode & OutlineRenderFlags.EnableAlphaTesting) != 0)
+			{
+				EditorGUILayout.PropertyField(cutoffProp, _cutoffContent);
+			}
+
+			serializedObject.ApplyModifiedProperties();
+		}
+
+		internal static float GetSettingsInstanceHeight(SerializedProperty property)
+		{
+			var lineCy = EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
+			var renderModeProp = property.FindPropertyRelative(_renderModePropName);
+			var renderMode = (OutlineRenderFlags)renderModeProp.intValue;
+			var result = lineCy * 4;
+
+			if ((renderMode & OutlineRenderFlags.Blurred) != 0)
+			{
+				result += lineCy;
+			}
+
+			if ((renderMode & OutlineRenderFlags.EnableAlphaTesting) != 0)
+			{
+				result += lineCy;
+			}
+
+			return result;
+		}
+
+		internal static float GetSettingsWithMaskHeight(SerializedProperty property)
+		{
+			var lineCy = EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
+			var layerMaskProp = property.FindPropertyRelative(_layerMaskPropName);
+
+			if (layerMaskProp.intValue != 0)
+			{
+				var renderModeProp = property.FindPropertyRelative(_renderModePropName);
+				var renderMode = (OutlineRenderFlags)renderModeProp.intValue;
+				var result = lineCy * 5;
+
+				if ((renderMode & OutlineRenderFlags.Blurred) != 0)
+				{
+					result += lineCy;
+				}
+
+				if ((renderMode & OutlineRenderFlags.EnableAlphaTesting) != 0)
+				{
+					result += lineCy;
+				}
+
+				return result;
+			}
+
+			return lineCy;
+		}
+
+		internal static void DrawSettingsInstance(Rect rc, SerializedProperty property)
+		{
+			var settingsProp = property.FindPropertyRelative(_settingsPropName);
+
+			EditorGUI.PropertyField(new Rect(rc.x, rc.y, rc.width, EditorGUIUtility.singleLineHeight), settingsProp);
+			EditorGUI.indentLevel += 1;
+
+			if (settingsProp.objectReferenceValue)
+			{
+				var obj = new SerializedObject(settingsProp.objectReferenceValue);
+				var colorProp = obj.FindProperty(_colorPropName);
+				var widthProp = obj.FindProperty(_widthPropName);
+				var intensityProp = obj.FindProperty(_intensityPropName);
+				var cutoffProp = obj.FindProperty(_cutoffPropName);
+				var renderModeProp = obj.FindProperty(_renderModePropName);
+
+				EditorGUI.BeginDisabledGroup(true);
+				DrawSettingsInternal(rc, colorProp, widthProp, intensityProp, cutoffProp, renderModeProp);
+				EditorGUI.EndDisabledGroup();
+			}
+			else
+			{
+				var colorProp = property.FindPropertyRelative(_colorPropName);
+				var widthProp = property.FindPropertyRelative(_widthPropName);
+				var intensityProp = property.FindPropertyRelative(_intensityPropName);
+				var cutoffProp = property.FindPropertyRelative(_cutoffPropName);
+				var renderModeProp = property.FindPropertyRelative(_renderModePropName);
+
+				DrawSettingsInternal(rc, colorProp, widthProp, intensityProp, cutoffProp, renderModeProp);
+			}
+
+			EditorGUI.indentLevel -= 1;
+		}
+
+		internal static void DrawSettingsWithMask(Rect rc, SerializedProperty property)
+		{
+			var layerMaskProp = property.FindPropertyRelative(_layerMaskPropName);
+
+			EditorGUI.PropertyField(new Rect(rc.x, rc.y, rc.width, EditorGUIUtility.singleLineHeight), layerMaskProp, _layerMaskContent);
+
+			if (layerMaskProp.intValue != 0)
+			{
+				var lineCy = EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
+				DrawSettingsInstance(new Rect(rc.x, rc.y + lineCy, rc.width, rc.height - lineCy), property);
 			}
 		}
 
-		private void OnEnable()
+		private static void DrawSettingsInternal(Rect rc, SerializedProperty colorProp, SerializedProperty widthProp, SerializedProperty intensityProp, SerializedProperty cutoffProp, SerializedProperty renderModeProp)
 		{
-			_settings = (OutlineSettings)target;
+			var renderMode = (OutlineRenderFlags)renderModeProp.intValue;
+			var lineCy = EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
+			var n = 4;
+
+			EditorGUI.PropertyField(new Rect(rc.x, rc.y + 1 * lineCy, rc.width, EditorGUIUtility.singleLineHeight), colorProp, _colorContent);
+			EditorGUI.PropertyField(new Rect(rc.x, rc.y + 2 * lineCy, rc.width, EditorGUIUtility.singleLineHeight), widthProp, _widthContent);
+
+			// NOTE: EditorGUI.PropertyField doesn't allow multi-selection, have to use EnumFlagsField explixitly.
+			renderModeProp.intValue = (int)(OutlineRenderFlags)EditorGUI.EnumFlagsField(new Rect(rc.x, rc.y + 3 * lineCy, rc.width, EditorGUIUtility.singleLineHeight), _renderModeContent, renderMode);
+
+			if ((renderMode & OutlineRenderFlags.Blurred) != 0)
+			{
+				EditorGUI.PropertyField(new Rect(rc.x, rc.y + n++ * lineCy, rc.width, EditorGUIUtility.singleLineHeight), intensityProp, _intensityContent);
+			}
+
+			if ((renderMode & OutlineRenderFlags.EnableAlphaTesting) != 0)
+			{
+				EditorGUI.PropertyField(new Rect(rc.x, rc.y + n * lineCy, rc.width, EditorGUIUtility.singleLineHeight), cutoffProp, _cutoffContent);
+			}
 		}
 	}
 }