+++ Outline.Core/Packages/UnityFx.Outline/Runtime/Scripts/OutlineLayerCollection.cs
@@ -23,6 +23,8 @@ namespace UnityFx.Outline
 		private List<OutlineLayer> _layers = new List<OutlineLayer>();
 		[SerializeField, HideInInspector]
 		private int _layerMask;
+		[SerializeField, HideInInspector]
+		private bool _mergeLayerObjects;
 
 		#endregion
 
@@ -48,6 +50,21 @@ namespace UnityFx.Outline
 						layer.UpdateRenderers(value);
 					}
 				}
+			}
+		}
+
+		/// <summary>
+		/// Gets or sets a value indicating whether layer game objects should be trated as one.
+		/// </summary>
+		public bool MergeLayerObjects
+		{
+			get
+			{
+				return _mergeLayerObjects;
+			}
+			set
+			{
+				_mergeLayerObjects = value;
 			}
 		}
 
@@ -102,9 +119,34 @@ namespace UnityFx.Outline
 		/// </summary>
 		public void GetRenderObjects(IList<OutlineRenderObject> renderObjects)
 		{
+			if (_mergeLayerObjects)
+			{
+				foreach (var layer in _layers)
+				{
+					renderObjects.Add(new OutlineRenderObject(layer.GetRenderers(), layer, layer.Name));
+				}
+			}
+			else
+			{
+				foreach (var layer in _layers)
+				{
+					layer.GetRenderObjects(renderObjects);
+				}
+			}
+		}
+
+		/// <summary>
+		/// Removes the specified <see cref="GameObject"/> from layers.
+		/// </summary>
+		/// <param name="go">A <see cref="GameObject"/> to remove.</param>
+		public void Remove(GameObject go)
+		{
 			foreach (var layer in _layers)
 			{
-				layer.GetRenderObjects(renderObjects);
+				if (layer.Remove(go))
+				{
+					break;
+				}
 			}
 		}